+++ src/TBoardView.h
@@ -20,8 +20,8 @@ public:
 public:
 	virtual void CairoDrawing();
 
-    int squareSize;
-    int left_margin,top_margin;
+	double squareSize;
+	double left_margin,top_margin;
 
     TLinesBoard* linesBoard;
